github_analyzer/requirements.txt
pandas
openpyxl
requests
PyGithub
nbformat
python-dotenv
groq
